fix: rename command
