+++ pkg/cli/experiment/command.go
@@ -3,7 +3,7 @@ package experiment
 import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
-	"github.com/suzuki-shunsuke/ghalint/pkg/cli/experiment/schema"
+	"github.com/suzuki-shunsuke/ghalint/pkg/cli/experiment/validateinput"
 	"github.com/urfave/cli/v3"
 )
 
@@ -14,7 +14,7 @@ func New(logE *logrus.Entry, fs afero.Fs) *cli.Command {
 		Usage:       "experimental commands",
 		Description: "experimental commands. These commands are not stable and may change in the future without major updates.",
 		Commands: []*cli.Command{
-			schema.New(logE, fs),
+			validateinput.New(logE, fs),
 		},
 	}
 }
